+++ apps/web/drizzle/migrations/0011_superb_shockwave.sql
@@ -1,0 +1,16 @@
+PRAGMA foreign_keys=OFF;--> statement-breakpoint
+CREATE TABLE `__new_player` (
+	`id` text PRIMARY KEY NOT NULL,
+	`display_name` text,
+	`user_id` text,
+	`notes` text,
+	`created_at` integer NOT NULL,
+	`updated_at` integer NOT NULL,
+	FOREIGN KEY (`user_id`) REFERENCES `user`(`id`) ON UPDATE no action ON DELETE set null
+);
+--> statement-breakpoint
+INSERT INTO `__new_player`("id", "display_name", "user_id", "notes", "created_at", "updated_at") SELECT "id", "display_name", "user_id", "notes", "created_at", "updated_at" FROM `player`;--> statement-breakpoint
+DROP TABLE `player`;--> statement-breakpoint
+ALTER TABLE `__new_player` RENAME TO `player`;--> statement-breakpoint
+PRAGMA foreign_keys=ON;--> statement-breakpoint
+CREATE INDEX `player_user_id_idx` ON `player` (`user_id`);